added confirmation modal before deleting a person and before logout

+++ frontend-app/src/app/home-page/home-page.component.ts
@@ -1,8 +1,10 @@
-import { Component } from '@angular/core';
+import { Component, ViewChild } from '@angular/core';
 import { PeopleService } from '../services/people.service';
 import { PersonOverview } from '../domain/person-overview';
 import { InterceptorService } from '../services/interceptor.service';
 import { ToastService } from '../services/toast.service';
+import { ModalComponent } from '../modal/modal.component';
+import { ModalConfig } from '../domain/modal-config';
 
 @Component({
   selector: 'app-home-page',
@@ -11,7 +13,17 @@ import { ToastService } from '../services/toast.service';
 })
 export class HomePageComponent {
 
+  @ViewChild('modal') private modalComponent: ModalComponent | undefined;
+  modalConfig: ModalConfig = {
+    modalTitle: 'Delete person?',
+    dismissButtonLabel: 'Cancel',
+    closeButtonLabel: 'Confirm',
+    onDismiss: () => false,
+    onClose: () => true,
+  } as ModalConfig;
+
   peopleList: PersonOverview[] = [];
+  toDelete: PersonOverview | undefined;
   role: string = "";
 
   constructor(private peopleService: PeopleService,
@@ -27,10 +39,17 @@ export class HomePageComponent {
     });
   }
 
-  deletePerson(id: number) {
-    this.peopleService.deletePerson(id).subscribe(res => {
-      this.peopleList = this.peopleList.filter(p => p.id != id);
-      this.toastService.showSuccess(`Successfully deleted person ${res.firstName} ${res.lastName}`);
-    })
+  async deletePerson(id: number) {
+    this.toDelete = this.peopleList.find(p => p.id == id);
+    if(await this.openModal()) {
+      this.peopleService.deletePerson(id).subscribe(res => {
+        this.peopleList = this.peopleList.filter(p => p.id != id);
+        this.toastService.showSuccess(`Successfully deleted person ${res.firstName} ${res.lastName}`);
+      })
+    }
+  }
+
+  async openModal() {
+    return await this.modalComponent?.open()
   }
 }
